pyproject.toml
[tool.poetry]
name = "booster-school"
version = "0.1.0"
description = ""
authors = ["WitchInTheHandkerchief <ilimkanu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.1"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.9"
django-environ = "^0.11.2"
black = "^24.1.1"
flake8 = "^7.0.0"
django-pytest = "^0.2.0"
django-cors-headers = "^4.3.1"
pillow = "^10.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
